\name{ao.asym}
\alias{ao.asym}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Defining the link-glm object for the AO asymmetric transformation 
}
\description{
This function defines the AO asymmetric link function used by glm() and glm.fit() for the binomial family
}
\usage{
ao.asym(phi, verbose = FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{phi}{
Transformation parameter
}
  \item{verbose}{
Verbose message
}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
