\name{ao.glm.default}
\alias{ao.glm.default}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Default method 
}
\description{
Default method of ao.glm()
}
\usage{
\method{ao.glm}{default}(x, y, link, phi, weights, ...)
%ao.glm.default(x, y, link, phi, weights, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Design matrix
}
  \item{y}{
Response variable
}
  \item{link}{
Link function: ao.sym or ao.asym
}
 \item{phi}{
An optional vector specifying the value(s) of the transformation parameter to
}
  \item{weights}{
An optional vector of weights to be used in the fitting process
}
  \item{\dots}{
Further arguments passed to or from other methods
}
}
\author{
Hakim-Moulay Dehbi
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
