\name{AOfamilies-package}
\alias{AOfamilies-package}
\alias{AOfamilies}
\docType{package}
\title{
Fitting GLMs and QR models using AO transformations 
}
\description{
This package implements both the AO symmetric and asymmetric transformations for binary response variables in GLM and for bounded response variables in QR models. The MLE of the transformation parameter is obtained using the profile log-likelihood method
}
\details{
\tabular{ll}{
Package: \tab AOfamilies\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-10-16\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Hakim-Moulay Dehbi (with contributions from Mario Cortina-Borja and Marco Geraci)

Maintainer: Dehbi, Hakim-Moulay <hakim-moulay.dehbi12@imperial.ac.uk>

}
\references{
Aranda-Ordaz, F. J. (1981). On two families of transformations to additivity for binary response data. Biometrika 68 (2), pp. 357-363

Bottai, M., B. Cai, and R. E. McKeown (2010). Logistic quantile regression for bounded outcomes. Statistics in Medicine 29 (2), 309-317

For the datasets used in this manual, see: 

1/ Bliss, C. I. (1935). The calculation of the dosage-mortality curve. Annals of Applied Biology 22 (1), 134-167

2/ Davis, C. S. (1991). Semi-parametric and non-parametric methods for the analysis of repeated measurements with applications to clinical trials. Statistics in Medicine 10 (12), 1959-1980

3/ Hand et al. (1993). A Handbook of Small Data Sets. Taylor & Francis
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ package }
