\name{print.ao.qr}
\alias{print.ao.qr}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Printing the MLE, its associated log-likelihood and the coefficients of an "ao.qr" object}
\description{
Print method of the function \code{\link{ao.qr}}
}
\usage{
\method{print}{ao.qr} (x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
An object of class "ao.qr"
}
  \item{...}{
Not used
}
}
\details{
Printing the MLE of the transformation parameter, its associated log-likelihood and the coefficients of the regression
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
