\name{ao.sym}
\alias{ao.sym}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
AO symmetric transformation
}
\description{
Function to define the AO symmetric transformation
}
\usage{
ao.sym(phi, verbose = FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{phi}{
Transformation parameter
}
  \item{verbose}{
TRUE or FALSE. Default is FALSE
}
}
\author{
Hakim-Moulay Dehbi
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
